backend/pyproject.toml
[tool.poetry]
name = "textbook-rag-backend"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95"
uvicorn = "^0.22"
python-multipart = "*"
sentence-transformers = "*"
faiss-cpu = "*"
PyPDF2 = "*"
openai = "*"
pydantic = "*"
aiofiles = "*"
